Rails Rubocop tweaks.
In some cases, these changes make sense (using join, and if+blank). In others, the rules feel like warnings ('this may not do what you want it to!'), and yet, they are the precise behaviour I'm after. Hence, disabling the rules within the source - though this feels like an inelegant solution.